Minor fixups for OSX, esp. making the NEW_BUFFER standard -- not an option.

+++ portmidi/trunk/pm_common/pminternal.h
@@ -135,7 +135,7 @@ typedef struct pm_internal_struct {
      * important
      */
     unsigned char *fill_base; /* addr of ptr to sysex data */
-    int *fill_offset_ptr; /* offset of next sysex byte */
+    unsigned long *fill_offset_ptr; /* offset of next sysex byte */
     int fill_length; /* how many sysex bytes to write */
 } PmInternal;
